Reorganized file structure
Placed every page related to the dashboard in its separate folder inside the Dashboard folder.
